changed unsigned int -> size_t to fix segfault on 64-bit machine
git-svn-id: svn+ssh://login2/data/avl11/svnroot/MOOS/trunk@2364 5f621d73-a7ea-4b79-bc62-7e07601040e3

+++ Tools/Graphical/uPlayBack/MOOSPlayBackV2.cpp
@@ -263,7 +263,7 @@ bool CMOOSPlayBackV2::MessageFromLine(const std::string & sLine, CMOOSMsg &Msg)
     m_ALog.GetNextToken(sLine,n,sKey);
     m_ALog.GetNextToken(sLine,n,sSrc);
 
-    int nData = sLine.find_first_not_of(" \t",n);
+    size_t nData = sLine.find_first_not_of(" \t",n);
     if(nData==string::npos)
         return false;
